Fixed up AI:
Will no longer end on a tile with another enemy. This was happening when there was an enemy on a tile next to a player and another enemy passed over that tile and took an attack of opportunity causing it to die. The code was setting the enemy that tile contained to null, but will now only nullify it if it is the enemy which died.

Enemies will now try to avoid taking attacks of opportunity.

Enemies will now register the tile they are standing on currenlty as a valid location to end their movement.

Powered up the enemies a bit.